Factor out shrinker into its own file
