+++ hypothesis-python/RELEASE.rst
@@ -1,0 +1,4 @@
+RELEASE_TYPE: patch
+
+This release is a pure refactoring that extracts some internal code into its own file.
+It should have no user visible effect.
